server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
modal==0.63.96
aiohttp==3.9.3
pydantic==2.6.0
